requirements.txt
streamlit >= 1.28.0
streamlit-chat >= 0.1.1
streamlit-option-menu >= 0.3.6
groq >= 0.4.0
chromadb >= 0.4.15
sentence-transformers >= 2.2.2
transformers >= 4.35.0
pandas >= 2.0.0
numpy >= 1.24.0
scikit-learn >= 1.3.0
streamlit-session-state >= 0.1.0
streamlit-elements >= 0.1.0
streamlit-card >= 0.0.5
streamlit-pills >= 0.3.0
plotly >= 5.17.0
matplotlib >= 3.7.0
seaborn >= 0.12.0
reportlab >= 4.0.0
fpdf2 >= 2.7.0
weasyprint >= 60.0
streamlit-pdf >= 0.0.1
streamlit-pydantic >= 0.6.0
pydantic >= 2.4.0
transitions >= 0.9.0
python-dateutil >= 2.8.0
python-dotenv >= 1.0.0
pydantic-settings >= 2.0.0
streamlit-authenticator >= 0.2.3
streamlit-feedback >= 0.1.0
streamlit-analytics >= 0.4.0
loguru >= 0.7.0
